Add files via upload
The aim of this project was to create a teaching tool to help students assemble MIRTO robot. The application was built for Hololens headset, but can also be used for Google cardboard headset, which is more affordable for university, to equip for this usage. Students will be following instructions in Augmented Reality instead of traditional paper handouts. I believe this will be a great opportunity for universities to bring the classes to different level of studies and build strong interest for students. The project will be evaluated by allowing students to use the application and collect feedback in regards to the using AR versus the use of paper handouts. Possible future improvements, will be provided in the conclusion section.